Add Patient Refund Policy

+++ BackEnd/DoctorAppointmentsSystem/Services.Abstraction/ITransactionService.cs
@@ -8,7 +8,7 @@ namespace Services.Abstraction
         Task<string> GetPaymentId(int transactionId);
         Task<int> GetByPaymentId(string paymentId);
         Task AddAsync(int patientId, int doctorId, int amount, string paymentId);
-        Task UpdateAsync(int id, int amount);
+        Task UpdateAsync(int id, double amount);
         Task DeleteAsync(int id);
     }
 }
